+++ iModelBridgeCore/Imagepp/all/gra/hgf/src/HGF2DHoledShape.cpp
@@ -2,7 +2,7 @@
 //:>
 //:>     $Source: all/gra/hgf/src/HGF2DHoledShape.cpp $
 //:>
-//:>  $Copyright: (c) 2013 Bentley Systems, Incorporated. All rights reserved. $
+//:>  $Copyright: (c) 2015 Bentley Systems, Incorporated. All rights reserved. $
 //:>
 //:>+--------------------------------------------------------------------------------------
 // Methods for class HGF2DHoledShape
@@ -1920,7 +1920,7 @@ void HGF2DHoledShape::PrintState(ostream& po_rOutput) const
     HDUMP0("Object is a HGF2DHoledShape\n");
 
     po_rOutput << "The holed shape contains " << m_HoleList.size() << "holes" << endl;
-    HDUMP1("The holed shape contains %ld holes", m_HoleList.size());
+    HDUMP1("The holed shape contains %lld holes", (uint64_t)m_HoleList.size());
 
     m_pBaseShape->PrintState(po_rOutput);
